Added dictionary methods to the classes and added functions to retrieve dashboard related data

+++ kame/kame_dashboard/views.py
@@ -1,56 +1,58 @@
 from django.http import JsonResponse
-from kame_app.models import Patient, CustomUser
+from kame_app.models import Patient, CustomUser, MedicalRecord
+import json
+from django.core import serializers
+
 
 def dashboard(request):
-    # Fetch all patients with related medical records and ICD10 codes
-    patients = Patient.objects.all().prefetch_related('medicalrecord_set__icd10')
-
-    # Prepare data to be serialized
-    data = []
-    for patient in patients:
-        records = []
-        for record in patient.medicalrecord_set.all():
-            records.append({
-                'description': record.description,
-                'icd10_code': record.icd10.code,  # Assuming 'code' is a field in the Icd10 model
-                'icd10_description': record.icd10.description  # Assuming 'description' is a field in the Icd10 model
-            })
-        data.append({
-            'full_name': patient.full_name,
-            'age': patient.age,
-            'gender': patient.gender,
-            'phone_number': patient.phone_number,
-            'address': patient.address,
-            'city': patient.city,
-            'country': patient.country,
-            'date_of_birth': patient.date_of_birth.isoformat(),
-            'ssn': patient.ssn,
-            'medical_records': records,
-        })
-
-    # Return the data as a JSON response
-    return JsonResponse(data, safe=False)
+    data = MedicalRecord.objects.all()
+    return JsonResponse([record.to_dict() for record in data], safe=False)
 
 def users(request):
-    data = CustomUser.objects.all()
-    return JsonResponse(data, safe=False)
+    if request.method == "GET":
+        data = CustomUser.objects.all()
+        return JsonResponse([user.to_dict() for user in data], safe=False)
+    else:
+        return JsonResponse({'error': 'Invalid request method'}, status=405)
 
 def get_user(request, employee_id):
-    data = CustomUser.objects.get(pk=employee_id)
-    return JsonResponse(data, safe=False)
+    if request.method == 'GET':
+        data = CustomUser.objects.get(pk=employee_id)
+        return JsonResponse(data, safe=False)
 
 def update_user(request, employee_id):
-    data = CustomUser.objects.get(pk=employee_id)
-    return JsonResponse(data, safe=False)
+    if request.method == "PUT":
+
+        data = CustomUser.objects.get(pk=employee_id)
+        return JsonResponse(data, safe=False)
 
 def delete_user(request, employee_id):
-    try:
-        user = CustomUser.objects.get(pk=employee_id)
-        user.delete()
-        return JsonResponse({'message': 'User deleted succesfully'}, status=200)
-    except CustomUser.DoesNotExist:
-        return JsonResponse({'error': 'User not found'}, status=404)
+    if request.method == "DELETE":
+        try:
+            user = CustomUser.objects.get(pk=employee_id)
+            user.delete()
+            return JsonResponse({'message': 'User deleted succesfully'}, status=200)
+        except CustomUser.DoesNotExist:
+            return JsonResponse({'error': 'User not found'}, status=404)
 
-def get_patinets(request):
+def get_patients(request):
     data = Patient.objects.all()
     return JsonResponse(data, safe=False)
+
+def create_user(request):
+    if request.method == 'POST':
+        try:
+            data = json.loads(request.body)
+            role = data.get('role')
+            email = data.get('email')
+            password = data.get('password')
+            name = data.get('name')
+
+            if not role or not email or not password or not name:
+                return JsonResponse({'error': 'Missing required fields'}, status=400)
+            user = CustomUser.objects.create_user(role=role, name=name, email=email, password=password)
+            return JsonResponse({'message': 'User created successfully', 'user_id': user.id}, status=201)
+        except Exception as e:
+            return JsonResponse({'error': str(e)}, status=500)
+    else:
+        return JsonResponse({'error': 'Invalid request method'}, status=405)
